Add `qmk find` command, abstract searching out to separate file.
